+++ tests/testthat/test-delete_values.R
@@ -2,7 +2,7 @@
 test_that("delete_values() works (basic test)", {
   expect_equal(
     count_NA(delete_values(
-      "MAR", "rank",
+      "MAR_rank",
       ds = df_XY_20, p = 0.1,
       cols_mis = "X", cols_ctrl = "Y", n_mis_stochastic = FALSE
     )),
@@ -35,6 +35,7 @@ test_that("delete_values() deprecate ctrl_cols", {
     c(X = 2, Y = 0)
   )
 })
+
 
 test_that("delete_values() deprecate stochastic", {
   # use delete_MCAR() as calling function (arbitrary choice)
@@ -80,11 +81,22 @@ test_that("delete_values() calls check_args_MNAR()", {
 test_that("delete_values() adjusts p", {
   expect_equal(
     count_NA(delete_values(
-      "MAR", "rank",
+      "MAR_rank",
       ds = df_XYZ_100, p = 0.1,
       cols_mis = c("X", "Z"), cols_ctrl = c("Y", "Y"), n_mis_stochastic = FALSE
     )),
     c(X = 10, Y = 0, Z = 10)
+  )
+})
+
+
+test_that("delete_values() errors for wrong mech_type", {
+  expect_error(
+    delete_values("not_existing_mech",
+      ds = df_XYZ_100, p = 0.1, cols_mis = 1,
+      n_mis_stochastic = FALSE
+    ),
+    "Invalid missing data mechanismus 'not_existing_mech'"
   )
 })